pyproject.toml
[project]
name = "pyxora"
version = "0.6.0.dev0"
authors = [{ name = "AntonisDevStuff", email = "contact@antonisdevstuff.com" }]
description = "A game engine library using pygame-ce, pymunk, and more"
keywords = ["engine", "game engine", "game development"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce==2.5.6",
    "pymunk==7.2.0",
    "pdoc==16.0.0",
    "cx_Freeze==8.4.1",
    "pygbag==0.9.2",
    "pillow==12.0.0",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: pygame",
]
[project.urls]
Homepage = "https://pyxora.github.io/pyxora-website"
Documentation = "https://pyxora.github.io/pyxora-docs"
Repository = "https://github.com/pyxora/pyxora"
Issues = "https://github.com/pyxora/pyxora/issues"
Releases = "https://github.com/pyxora/pyxora/releases"

[project.scripts]
pyxora = "pyxora.__main__:main"

[tool.setuptools]
package-dir = { "pyxora" = "src" }
packages = [
    "pyxora",
    "pyxora.wrapper",
    "pyxora.utils",
    "pyxora.docs",
    "pyxora.data",
    "pyxora.projects"
]

[tool.setuptools.package-data]
pyxora = ["data/**"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
